core/api/api_views.py
from rest_framework import viewsets
from core.models import Wallet, Transaction, SavingGoal
from core.api.serializers import WalletSerializer, TransactionSerializer, SavingGoalSerializer


class WalletViewSet(viewsets.ModelViewSet):
    queryset = Wallet.objects.all()
    serializer_class = WalletSerializer

class TransactionViewSet(viewsets.ModelViewSet):
    queryset = Transaction.objects.all()
    serializer_class = TransactionSerializer

class SavingGoalViewSet(viewsets.ModelViewSet):
    queryset = SavingGoal.objects.all()
    serializer_class = SavingGoalSerializer


